Rust::Com Optimize the init FFI calls
* Move the init in FFI trait
* Created common_rs for ffi
